move modelGL to Gem/
